requirements.txt
```text
google-genai
argparse